feat: add logs in ProductService

+++ src/main/java/org/example/ecommerce/service/ProductService.java
@@ -1,5 +1,7 @@
 package org.example.ecommerce.service;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.example.ecommerce.dto.request.ProductRequestDTO;
 import org.example.ecommerce.dto.response.ProductResponseDTO;
 import org.example.ecommerce.model.Product;
@@ -16,23 +18,46 @@ public class ProductService {
     @Autowired
     ProductRepository productRepository;
 
+    private static final Logger LOGGER = LogManager.getLogger();
+
     public ProductResponseDTO findById(UUID id) {
+        LOGGER.info("Getting product...");
         return productRepository.findById(id)
                 .map(ProductResponseDTO::new)
-                .orElseThrow(() -> new RuntimeException("Id not found."));
+                .orElseThrow(() -> {
+                    LOGGER.error("Product not found.");
+                    return new RuntimeException("Product not found.");
+                });
     }
 
     public ProductResponseDTO findByName(String name) {
         return productRepository.findByName(name)
                 .map(ProductResponseDTO::new)
-                .orElseThrow(() -> new RuntimeException("Product not found."));
+                .orElseThrow(() -> {
+                    LOGGER.error("Product not found.");
+                    return  new RuntimeException("Product not found.");
+                });
     }
 
     public List<ProductResponseDTO> findByStore(String soldBy) {
+        LOGGER.info("Getting products by store...");
+
+        if(productRepository.findBySoldBy(soldBy).isEmpty()) {
+            LOGGER.error("The products List is empty");
+            throw new RuntimeException("The products list is empty");
+        }
+
         return productRepository.findBySoldBy(soldBy).stream().map(ProductResponseDTO::new).toList();
     }
 
     public List<ProductResponseDTO> findAllProducts() {
+        LOGGER.info("Getting all products...");
+
+        if(productRepository.findAll().isEmpty()) {
+            LOGGER.error("The products List is empty");
+            throw new RuntimeException("The products list is empty");
+        }
+
         return productRepository.findAll().stream().map(ProductResponseDTO::new).toList();
     }
 
@@ -54,6 +79,14 @@ public class ProductService {
     }
 
     public void deleteProduct(UUID id) {
+        LOGGER.info("Deleting product...");
+
+        if (productRepository.findById(id).isEmpty()) {
+            LOGGER.error("Product not found.");
+            throw new RuntimeException("Product not found.");
+        }
+
         productRepository.deleteById(id);
+        LOGGER.info("Product deleted successfully.");
     }
 }
